+++ operators/__init__.py
@@ -57,7 +57,8 @@ from .select_strips_under_cursor import POWER_SEQUENCER_OT_select_strips_under_c
 from .markers_set_preview_in_between import POWER_SEQUENCER_OT_set_preview_between_markers
 from .set_timeline_range import POWER_SEQUENCER_OT_set_timeline_range
 from .trim_left_or_right_handles import POWER_SEQUENCER_OT_trim_left_or_right_handles
-from .snap_selection_to_cursor import POWER_SEQUENCER_OT_snap_selection_to_cursor
+from .snap import POWER_SEQUENCER_OT_snap
+from .snap_selection import POWER_SEQUENCER_OT_snap_selection
 from .speed_remove_effect import POWER_SEQUENCER_OT_unspeed
 from .swap_strips import POWER_SEQUENCER_OT_swap_strips
 from .synchronize_titles import POWER_SEQUENCER_OT_synchronize_titles
@@ -126,7 +127,8 @@ classes = [
     POWER_SEQUENCER_OT_set_preview_between_markers,
     POWER_SEQUENCER_OT_set_timeline_range,
     POWER_SEQUENCER_OT_trim_left_or_right_handles,
-    POWER_SEQUENCER_OT_snap_selection_to_cursor,
+    POWER_SEQUENCER_OT_snap,
+    POWER_SEQUENCER_OT_snap_selection,
     POWER_SEQUENCER_OT_unspeed,
     POWER_SEQUENCER_OT_swap_strips,
     POWER_SEQUENCER_OT_synchronize_titles,