+++ operators/select_strips_under_cursor.py
@@ -1,5 +1,6 @@
 import bpy
 
+from .utils.functions import get_sequences_under_cursor
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
 
 
@@ -20,26 +21,15 @@ class POWER_SEQUENCER_OT_select_strips_under_cursor(bpy.types.Operator):
     bl_description = doc_brief(doc["description"])
     bl_options = {"REGISTER", "UNDO"}
 
-    locked: bpy.props.BoolProperty(name="Locked")
+    deselect_first: bpy.props.BoolProperty(name="Deselect First")
 
     @classmethod
     def poll(cls, context):
         return len(context.sequences) > 0
 
     def execute(self, context):
-        bpy.ops.sequencer.select_all(action="DESELECT")
-        current_frame = context.scene.frame_current
-        sequences_to_select = []
-        for s in context.sequences:
-            if (
-                s.frame_final_start <= current_frame
-                and s.frame_final_end >= current_frame
-                and not s.lock
-                or self.locked
-            ):
-                sequences_to_select.append(s)
-        if not sequences_to_select:
-            return {"CANCELLED"}
-        for s in sequences_to_select:
+        if self.deselect_first:
+            bpy.ops.sequencer.select_all(action="DESELECT")
+        for s in get_sequences_under_cursor(context):
             s.select = True
         return {"FINISHED"}