+++ ui/menu_contextual.py
@@ -54,7 +54,7 @@ class POWER_SEQUENCER_MT_contextual(bpy.types.Menu):
                 "power_sequencer.ripple_delete", icon="AUTOMERGE_ON", text="Ripple delete"
             )
             layout.operator(
-                "power_sequencer.snap_selection_to_cursor", icon="SNAP_ON", text="Snap selection"
+                "power_sequencer.snap_selection", icon="SNAP_ON", text="Snap selection"
             )
 
         layout.separator()